+++ packages/piveau-ui-cli/src/schemas/index.ts
@@ -1,0 +1,11 @@
+// re-exports all schemas
+
+export * from './apiSchema'
+export * from './authenticationSchema'
+export * from './contentSchema'
+export * from './routingSchema'
+export * from './metadataSchema'
+export * from './languagesSchema'
+export * from './servicesSchema'
+export * from './themesSchema'
+export * from './trackerSchema'