+++ packages/piveau-ui-cli/src/user-config.js
@@ -1,0 +1,5 @@
+// user-config.js
+
+export const glueConfig = {
+  // ...
+}