+++ packages/piveau-ui-cli/src/user-config.ts
@@ -1,0 +1,6 @@
+// user-config.ts
+import { defineUserConfig } from '@piveau/piveau-hub-ui-modules'
+
+export const glueConfig = defineUserConfig({
+  // ...
+})